Rename converters according to the WPF default converters.
