+++ src/Soloplan.WhatsON.GUI.Common/Converters/CountToVisibilityConverter.cs
@@ -1,36 +1,37 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="InvertBoolConverter.cs" company="Soloplan GmbH">
+// <copyright file="CountVisibilityConverter.cs" company="Soloplan GmbH">
 //   Copyright (c) Soloplan GmbH. All rights reserved.
 //   Licensed under the MIT License. See License-file in the project root for license information.
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
-
 namespace Soloplan.WhatsON.GUI.Common.Converters
 {
   using System;
+  using System.Collections;
   using System.Globalization;
+  using System.Windows;
   using System.Windows.Data;
 
-  public class InvertBoolConverter : IValueConverter
+  /// <summary>
+  /// Converts count to visibility. Count ==  0 -> Hidden, Count > 0 -> Visible.
+  /// </summary>
+  internal class CountVisibilityConverter : IValueConverter
   {
+    public Visibility ValueForFalse { get; set; } = Visibility.Hidden;
+
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-      return Negate(value);
+      if (value is IList list)
+      {
+        return list.Count > 0 ? Visibility.Visible : this.ValueForFalse;
+      }
+
+      return this.ValueForFalse;
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
     {
-      return Negate(value);
-    }
-
-    private static object Negate(object value)
-    {
-      if (value is bool boolean)
-      {
-        return !boolean;
-      }
-
-      return false;
+      return null;
     }
   }
 }
