+++ src/Soloplan.WhatsON.Jenkins.GUI/JenkinsProjectDataTemplate.xaml
@@ -7,7 +7,7 @@
   <BooleanToVisibilityConverter x:Key="BoolToVisibility" />
   <converters:StatusToColorConverter x:Key="StatusToColorConverter" />
   <converters:TimeToApproximateTimeConverter x:Key="TimeToApproximateTimeConverter" />
-  <converters:NullOrDefaultVisibilityConverter x:Key="NullOrDefaultVisibilityConverter" />
+  <converters:NullOrDefaultToVisibilityConverter x:Key="NullOrDefaultVisibilityConverter" />
 
   <DataTemplate DataType="{x:Type gui:JenkinsStatusViewModel}">
     <buildServer:HistoryBuildControl DataContext="{Binding}">
